refactor: split monolithic index.ts into modular files
